AVI-HardSUB
Avi Hard Sub, is a video-hardsubber application running on Windows. This
application made using visual c# and of course AV Library. The input of
this application is video and text file, the format for text file is
similar to SRT. User can edit the SRT and hardsub the video using this
application.